+++ crates/dslab-mp/src/mc/node.rs
@@ -5,11 +5,11 @@ use std::rc::Rc;
 
 use crate::context::Context;
 use crate::mc::events::McEvent;
+use crate::mc::events::SystemTime;
 use crate::mc::network::McNetwork;
 use crate::message::Message;
 use crate::node::{EventLogEntry, ProcessEntry, ProcessEvent, TimerBehavior};
 use crate::process::ProcessState;
-use crate::mc::events::SystemTime;
 pub struct ProcessEntryState {
     pub proc_state: Box<dyn ProcessState>,
     pub event_log: Vec<EventLogEntry>,
@@ -129,7 +129,7 @@ impl McNode {
                         let event = McEvent::TimerFired {
                             timer: name.clone(),
                             proc: proc.clone(),
-                            duration: SystemTime::from(delay),
+                            timer_delay: SystemTime::from(delay),
                         };
                         new_events.push(event);
                         // event_id is 0 since it is not used in model checking
